+++ src/main/java/de/jeff_media/AngelChest/WorldListener.java
@@ -1,0 +1,19 @@
+package de.jeff_media.AngelChest;
+
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.world.WorldLoadEvent;
+
+public class WorldListener implements Listener {
+
+    Main main;
+
+    WorldListener(Main main) {
+        this.main=main;
+    }
+
+    @EventHandler
+    public void onWorldLoad(WorldLoadEvent e) {
+        main.loadAllAngelChestsFromFile();
+    }
+}